+++ src/features/GridScaling.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <Geode/binding/EditorUI.hpp>
+
+using namespace geode::prelude;
+
+void decrementGridSize(EditorUI* ui);
+void incrementGridSize(EditorUI* ui);